refactor: split region map political labels into province and sea parts (#3878)
Extract province-name and sea-zone label rendering from the monolithic
political labels part into dedicated part files for Phase 3 flame map
modularization.
