+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -30,7 +30,8 @@ part 'region_map_component_render_orchestrator.dart';
 part 'region_map_component_render_core.dart';
 part 'region_map_component_render_core_base_tiles.dart';
 part 'region_map_component_render_core_overlays.dart';
-part 'region_map_component_render_political_labels.dart';
+part 'region_map_component_render_political_labels_province.dart';
+part 'region_map_component_render_political_labels_sea.dart';
 part 'region_map_component_render_political_borders.dart';
 part 'region_map_component_render_markers.dart';
 part 'region_map_component_render_markers_settlements.dart';
